Extract healthCheck
